Former-commit-id: 30fbdb98c91ced1e2f56c5c4de5d7384a7d30a99

+++ Core/Main/OsmSharp.Osm.Routing/Interpreter/OsmRoutingInterpreter.cs
@@ -51,7 +51,7 @@ namespace OsmSharp.Osm.Routing.Interpreter
         /// <param name="along"></param>
         /// <param name="to"></param>
         /// <returns></returns>
-        public bool CanBeTraversed(uint from, uint along, uint to)
+        public bool CanBeTraversed(long from, long along, long to)
         {
             return true;
         }
